day 6 and 7 (partly done)

+++ 5/Part2.cs
@@ -3,7 +3,7 @@ namespace AOCDayTemplate
 {
     internal class Part2
     {
-        internal static void Solve(int[][][] input)
+        internal static void Solve(long[][][] input)
         {
             throw new NotImplementedException();
         }
